Move client-facing channel types into server crate
